feat: Cria contexto de transações usando useContext
- Compartilha entre os componentes: Summary e TransactionsTable
