+++ src/components/Summary.tsx/index.tsx
@@ -1,9 +1,14 @@
+import React, { useContext } from 'react';
 import { Container } from './styles';
 import incomeImg from '../../assets/income.svg';
 import outcomeImg from '../../assets/outcome.svg';
 import totalImg from '../../assets/total.svg';
+import { TransactionsContext } from '../../TransactionsContext';
 
 export function Summary() {
+    const transactions = useContext(TransactionsContext);
+    console.log(transactions);
+
     return (
         <Container>
             <div>
